requirements.txt
asgiref==3.6.0
Django==4.2.1
djangorestframework==3.14.0
exceptiongroup==1.1.1
flake8==6.0.0
iniconfig==2.0.0
Markdown==3.4.3
mccabe==0.7.0
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.3.1
pytest-django==4.5.2
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
